+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## 0.104.1 _2025-03-18_
 
 ### Improvements